// src/models/RouteEnum.ts
enum RouteEnum {
  Root = "/",
  Signup = "/signup",
  Authorize = "/Authorize",
  Login = "/login",
  Logout = "/logout",
  ApiSample = "/api-sample",
  UsersList = "/users-list",
  UsersEdit = "/users-edit",
  AddUser = "/users-add",
  UserDetails = "/users-details",
  Home = "/home",
  Link1 = "/link1",
  Link2 = "/link2",
  NotFound = "/not-found"
}

export default RouteEnum;